More tests and some refactorings
