+++ wasm-rpc-stubgen/test-data/caller-no-dep/wit/caller.wit
@@ -1,0 +1,4 @@
+package test:caller;
+
+world caller {
+}
